+++ src/tools.h
@@ -22,7 +22,7 @@ void randomstr(char *string, int size);
 void remove_newline(char* line);
 int endswith(const char *str, const char *suffix);
 char *replace_string(const char *source, const char *match, const char *replace);
-char *trim(char *str, int *res_size);
+char *trim(char *str, size_t *res_size);
 
 #ifdef __cplusplus
 }
